refactor(Str): Move Str to the Framework namespace
