+++ devTools/check_trailing_whitespaces.sh
@@ -27,7 +27,7 @@ files_with_trailing_spaces=$(
     git grep -In "\\s$" \
         ':!tests/autoloaded/*' \
         ':!tests/e2e/*' \
-        ':!tests/phpunit/StrTest.php' \
+        ':!tests/phpunit/Framework/StrTest.php' \
         ':!tests/phpunit/StringNormalizerTest.php' \
         ':!tests/phpunit/Fixtures/Files/phpunit/format-whitespace/original-phpunit.xml' \
         ':!tests/benchmark/Tracing/benchmark-source' \
