+++ src/Logger/Html/StrykerHtmlReportBuilder.php
@@ -48,6 +48,7 @@ use function current;
 use function implode;
 use function in_array;
 use Infection\AbstractTestFramework\Coverage\TestLocation;
+use Infection\Framework\Str;
 use Infection\Metrics\MetricsCalculator;
 use Infection\Metrics\ResultsCollector;
 use Infection\Mutant\DetectionStatus;
@@ -59,7 +60,6 @@ use Infection\Mutator\MutatorFactory;
 use Infection\Mutator\MutatorResolver;
 use Infection\Mutator\ProfileList;
 use Infection\Mutator\Removal\MethodCallRemoval;
-use Infection\Str;
 use function ltrim;
 use function md5;
 use const PHP_EOL;
