+++ src/Logger/JsonLogger.php
@@ -35,10 +35,10 @@ declare(strict_types=1);
 
 namespace Infection\Logger;
 
+use Infection\Framework\Str;
 use Infection\Metrics\MetricsCalculator;
 use Infection\Metrics\ResultsCollector;
 use Infection\Mutant\MutantExecutionResult;
-use Infection\Str;
 use function json_encode;
 use const JSON_THROW_ON_ERROR;
