+++ src/Logger/TextFileLogger.php
@@ -38,9 +38,9 @@ namespace Infection\Logger;
 use function array_map;
 use function explode;
 use function implode;
+use Infection\Framework\Str;
 use Infection\Metrics\ResultsCollector;
 use Infection\Mutant\MutantExecutionResult;
-use Infection\Str;
 use const PHP_EOL;
 use function sprintf;
 use function str_repeat;
