+++ tests/phpunit/Framework/StrTest.php
@@ -33,9 +33,10 @@
 
 declare(strict_types=1);
 
-namespace Infection\Tests;
+namespace Infection\Tests\Framework;
 
-use Infection\Str;
+use Infection\Framework\Str;
+use function Infection\Tests\normalizeLineReturn;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
